requirements.txt
FlagEmbedding>=1.2.10
torch>=2.3.0
numpy>=1.25.0
qdrant-client>=1.10.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
langchain-openai>=0.1.7
fastapi>=0.110.0
uvicorn>=0.23.0
pyyaml>=6.0
requests>=2.31.0
pydantic>=2.6.0
